+++ src/routes/_authenticated/passives/create/index.tsx
@@ -1,0 +1,6 @@
+import CreatePassive from '@/features/passives/create'
+import { createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute('/_authenticated/passives/create/')({
+  component: CreatePassive,
+})